+++ nebari/src/tree/versioned.rs
@@ -14,7 +14,7 @@ use super::{
     modify::Modification,
     read_chunk,
     serialization::BinarySerialization,
-    KeyEvaluation, KeyRange, PagedWriter, PAGE_SIZE,
+    KeyRange, PagedWriter, ScanEvaluation, PAGE_SIZE,
 };
 use crate::{
     chunk_cache::CacheEntry,
@@ -368,7 +368,7 @@ where
         cache: Option<&ChunkCache>,
     ) -> Result<(), Error>
     where
-        KeyEvaluator: FnMut(&ArcBytes<'static>) -> KeyEvaluation,
+        KeyEvaluator: FnMut(&ArcBytes<'static>) -> ScanEvaluation,
         KeyReader: FnMut(ArcBytes<'static>, ArcBytes<'static>) -> Result<(), Error>,
         Keys: Iterator<Item = &'keys [u8]>,
     {
@@ -427,8 +427,8 @@ where
         cache: Option<&ChunkCache>,
     ) -> Result<bool, AbortError<CallerError>>
     where
-        NodeEvaluator: FnMut(&ArcBytes<'static>, &Self::ReducedIndex, usize) -> bool,
-        KeyEvaluator: FnMut(&ArcBytes<'static>, &Self::Index) -> KeyEvaluation,
+        NodeEvaluator: FnMut(&ArcBytes<'static>, &Self::ReducedIndex, usize) -> ScanEvaluation,
+        KeyEvaluator: FnMut(&ArcBytes<'static>, &Self::Index) -> ScanEvaluation,
         KeyRangeBounds: RangeBounds<&'keys [u8]> + Debug + ?Sized,
         ScanDataCallback: FnMut(
             ArcBytes<'static>,